+++ components/ui/MermaidDiagramSkeleton.tsx
